requirements.txt
gsplat
numpy
pillow
torch
torchmetrics
torchvision
tqdm
